feat: add equity controller
