+++ src/application.module.ts
@@ -22,9 +22,11 @@ import { SwaggerConfigModule } from './swagger-config/swagger-config.module';
 import { CalendarModule } from './calendar/calendar.module';
 import { ActivityModule } from './activity/activity.module';
 import { WebhookModule } from './webhooks/webhook.module';
+import { EquityController } from './equity/equity.controller';
+import { EquityModule } from './equity/equity.module';
 
 @Module({
-  controllers: [ApplicationController],
+  controllers: [ApplicationController, EquityController],
   providers: [Logger, ApplicationService],
   imports: [
     ConfigModule.forRoot({
@@ -87,6 +89,7 @@ import { WebhookModule } from './webhooks/webhook.module';
     SwaggerConfigModule,
     ActivityModule,
     WebhookModule,
+    EquityModule,
   ],
 })
 export class ApplicationModule {
