+++ test/activity/controller.e2e-spec.ts
@@ -725,569 +725,4 @@ describe('ActivityController (e2e)', () => {
     });
   })
 
-  describe("GET /member_right", () => {
-    describe("Basic Positive Tests", () => {
-      let insertedMember;
-      let insertedActivity;
-      let insertedCategory;
-      let insertedActivityCategory;
-      let insertedActivitySession1;
-      let insertedActivityTicket1;
-      let insertedActivitySessionTicket1;
-      let insertedOrderLog;
-      let insertedProduct;
-      let insertedCurrency;
-      let insertedOrderProduct;
-  
-      beforeEach(async () => {
-        insertedMember = await createTestMember(manager, {
-          appId: app.id,
-          role: "app-owner",
-        });
-  
-        insertedActivity = await createTestActivity(manager, {
-          app: app,
-          organizer: insertedMember,
-          isPrivate: false, // scenario: 'holding' condition
-          publishedAt: new Date(), // scenario: 'holding' condition,
-          isParticipantsVisible: true
-        });
-  
-        insertedCategory = await createTestCategory(manager, {
-          appId: app.id,
-          class: "activity",
-        });
-  
-        insertedActivityCategory = await createTestActivityCategory(manager, {
-          activity: insertedActivity,
-          category: insertedCategory,
-        });
-  
-        insertedActivitySession1 = await createTestActivitySession(manager, {
-          activity: insertedActivity,
-          startedAt: new Date("2020-01-01T00:00:00Z"),
-          endedAt: new Date("2020-01-02T00:00:00Z"),
-        });
-  
-        insertedActivityTicket1 = await createTestActivityTicket(manager, {
-          activity: insertedActivity,
-          startedAt: new Date("2020-01-01T00:00:00Z"),
-          endedAt: new Date("2020-01-02T00:00:00Z"),
-        });
-  
-        insertedActivitySessionTicket1 = await createTestActivitySessionTicket(
-          manager,
-          {
-            activitySession: insertedActivitySession1,
-            activityTicket: insertedActivityTicket1,
-            activitySessionType: "offline",
-          },
-        );
-  
-        insertedOrderLog = await createTestOrderLog(manager, {
-          member: insertedMember,
-          appId: app.id,
-        });
-  
-        insertedProduct = await createTestProduct(manager, {
-          id: `ActivityTicket_${insertedActivityTicket1.id}`,
-          type: "ActivityTicket",
-          target: insertedActivityTicket1.id,
-        });
-  
-        insertedCurrency = await createTestCurrency(manager, {
-          id: "TWD",
-        });
-  
-        insertedOrderProduct = await createTestOrderProduct(manager, {
-          order: insertedOrderLog,
-          product: insertedProduct,
-          currency: insertedCurrency,
-          productId: insertedActivity.id,
-          options: {
-            from: `/activities/${insertedActivity.id}`,
-            currencyId: insertedCurrency.id,
-            currencyPrice: 2000,
-          },
-        });
-      });
-  
-      it("should successfully retrieve member rights for a valid activity ticket", async () => {
-        const jwtSecret = application
-          .get<ConfigService<{ HASURA_JWT_SECRET: string }>>(ConfigService)
-          .getOrThrow("HASURA_JWT_SECRET");
-  
-        const token = jwt.sign(
-          {
-            memberId: insertedMember.id,
-            permissions: [],
-          },
-          jwtSecret,
-        );
-  
-        const dto: FetchMemberRightActivityTicketDTO = {
-          activityTicketId: insertedActivityTicket1.id,
-          sessionId: "",
-        };
-  
-        const response = await request(application.getHttpServer())
-          .get(
-            `/activity/member_right?&activityTicketId=${dto.activityTicketId}${dto.sessionId ? `&sessionId=${dto.sessionId}` : ""}`,
-          )
-          .set("host", appHost.host)
-          .set("Authorization", `Bearer ${token}`)
-          .expect(200);
-  
-        expect(response.status).toBe(200);
-      });
-  
-      it("should return a valid JSON response that matches the expected schema", async () => {
-        const jwtSecret = application
-          .get<ConfigService<{ HASURA_JWT_SECRET: string }>>(ConfigService)
-          .getOrThrow("HASURA_JWT_SECRET");
-  
-        const token = jwt.sign(
-          {
-            memberId: insertedMember.id,
-            permissions: [],
-          },
-          jwtSecret,
-        );
-  
-        const dto: FetchMemberRightActivityTicketDTO = {
-          activityTicketId: insertedActivityTicket1.id,
-          sessionId: "",
-        };
-  
-        const { body: data } = await request(application.getHttpServer())
-          .get(
-            `/activity/member_right?&activityTicketId=${dto.activityTicketId}${dto.sessionId ? `&sessionId=${dto.sessionId}` : ""}`,
-          )
-          .set("host", appHost.host)
-          .set("Authorization", `Bearer ${token}`)
-          .expect(200);
-  
-        const activitySchema = Joi.object({
-          id: Joi.string().required(),
-          title: Joi.string().required(),
-          coverUrl: Joi.string().allow(null),
-          categories: Joi.array().items(Joi.object()).required(),
-          isParticipantsVisible: Joi.boolean().required()
-        });
-  
-        const sessionSchema = Joi.object({
-          id: Joi.string().required(),
-          startedAt: Joi.date().iso(),
-          endedAt: Joi.date().iso(),
-          location: Joi.string().allow(""),
-          description: Joi.string().allow(""),
-          threshold: Joi.allow(null),
-          onlineLink: Joi.string().allow(null),
-          title: Joi.string().required(),
-          maxAmount: Joi.object(),
-          participants: Joi.object(),
-          isEnrolled: Joi.boolean(),
-          type: Joi.string(),
-          attended: Joi.boolean().required(),
-        });
-  
-        const invoiceSchema = Joi.object({
-          name: Joi.string().required(),
-          email: Joi.string().required(),
-          phone: Joi.string().required(),
-          orderProductId: Joi.string().required(),
-        });
-  
-        const responseSchema = Joi.object({
-          id: Joi.string().required(),
-          activity: activitySchema,
-          sessions: Joi.array().items(sessionSchema),
-          invoice: invoiceSchema,
-        });
-  
-        const { error } = responseSchema.validate(data);
-        expect(error).toBeUndefined();
-        expect(data.id).toEqual(insertedActivityTicket1.id);
-        expect(data.activity).toEqual({
-          id: insertedActivity.id,
-          title: insertedActivity.title,
-          coverUrl: insertedActivity.coverUrl,
-          categories: [
-            {
-              id: insertedCategory.id,
-              name: insertedCategory.name,
-            },
-          ],
-          isParticipantsVisible: true
-        });
-        expect(data.sessions).toEqual([
-          {
-            id: insertedActivitySession1.id,
-            startedAt: "2020-01-01T00:00:00.000Z",
-            endedAt: "2020-01-02T00:00:00.000Z",
-            location: insertedActivitySession1.location,
-            description: insertedActivitySession1.description,
-            threshold: insertedActivitySession1.threshold,
-            onlineLink: insertedActivitySession1.onlineLink,
-            title: insertedActivitySession1.title,
-            maxAmount: {
-              offline: 1,
-              online: 0,
-            },
-            participants: {
-              offline: 1,
-              online: 0,
-            },
-            isEnrolled: true,
-            type: insertedActivitySessionTicket1.activitySessionType,
-            attended: false,
-          },
-        ]);
-  
-        expect(data.invoice).toEqual({
-          name: insertedOrderLog.invoiceOptions.name,
-          email: insertedOrderLog.invoiceOptions.email,
-          phone: insertedOrderLog.invoiceOptions.phone,
-          orderProductId: insertedOrderProduct.id,
-        });
-      });
-  
-      it("should include all associated sessions in the response", async () => {
-        const insertedActivitySession2 = await createTestActivitySession(
-          manager,
-          {
-            activity: insertedActivity,
-            startedAt: new Date("2020-01-01T00:00:00Z"),
-            endedAt: new Date("2020-01-02T00:00:00Z"),
-          },
-        );
-  
-        const insertedActivitySessionTicket2 =
-          await createTestActivitySessionTicket(manager, {
-            activitySession: insertedActivitySession2,
-            activityTicket: insertedActivityTicket1,
-            activitySessionType: "offline",
-          });
-  
-        const jwtSecret = application
-          .get<ConfigService<{ HASURA_JWT_SECRET: string }>>(ConfigService)
-          .getOrThrow("HASURA_JWT_SECRET");
-  
-        const token = jwt.sign(
-          {
-            memberId: insertedMember.id,
-            permissions: [],
-          },
-          jwtSecret,
-        );
-  
-        const dto: FetchMemberRightActivityTicketDTO = {
-          activityTicketId: insertedActivityTicket1.id,
-          sessionId: "",
-        };
-  
-        const { body: data } = await request(application.getHttpServer())
-          .get(
-            `/activity/member_right?&activityTicketId=${dto.activityTicketId}${dto.sessionId ? `&sessionId=${dto.sessionId}` : ""}`,
-          )
-          .set("host", appHost.host)
-          .set("Authorization", `Bearer ${token}`)
-          .expect(200);
-  
-        const sessionSchema = Joi.array().items(
-          Joi.object({
-            id: Joi.string().required(),
-            startedAt: Joi.date().iso(),
-            endedAt: Joi.date().iso(),
-            location: Joi.string().allow(""),
-            description: Joi.string().allow(""),
-            threshold: Joi.allow(null),
-            onlineLink: Joi.string().allow(null),
-            title: Joi.string().required(),
-            maxAmount: Joi.object(),
-            participants: Joi.object(),
-            isEnrolled: Joi.boolean(),
-            type: Joi.string(),
-            attended: Joi.boolean().required(),
-          }),
-        );
-  
-        const { error } = sessionSchema.validate(data.sessions);
-        expect(error).toBeUndefined();
-        expect(data.sessions.length).toEqual(2);
-      });
-
-      it("should include all associated categories in the activity response", async () => {
-        const insertedCategory2 = await createTestCategory(manager, {
-          appId: app.id,
-          class: "activity",
-        });
-  
-        const insertedActivityCategory = await createTestActivityCategory(manager, {
-          activity: insertedActivity,
-          category: insertedCategory2,
-        });
-
-        const jwtSecret = application
-          .get<ConfigService<{ HASURA_JWT_SECRET: string }>>(ConfigService)
-          .getOrThrow("HASURA_JWT_SECRET");
-  
-        const token = jwt.sign(
-          {
-            memberId: insertedMember.id,
-            permissions: [],
-          },
-          jwtSecret,
-        );
-  
-        const dto: FetchMemberRightActivityTicketDTO = {
-          activityTicketId: insertedActivityTicket1.id,
-          sessionId: "",
-        };
-  
-        const { body: data } = await request(application.getHttpServer())
-          .get(
-            `/activity/member_right?&activityTicketId=${dto.activityTicketId}${dto.sessionId ? `&sessionId=${dto.sessionId}` : ""}`,
-          )
-          .set("host", appHost.host)
-          .set("Authorization", `Bearer ${token}`)
-          .expect(200);
-  
-
-        expect(data.activity.categories.length).toEqual(2);
-      });
-  
-      it("should apply filters and return only the specified session in the response", async () => {
-        const insertedActivitySession2 = await createTestActivitySession(
-          manager,
-          {
-            activity: insertedActivity,
-            startedAt: new Date("2020-01-01T00:00:00Z"),
-            endedAt: new Date("2020-01-02T00:00:00Z"),
-          },
-        );
-  
-        const insertedActivitySessionTicket2 =
-          await createTestActivitySessionTicket(manager, {
-            activitySession: insertedActivitySession2,
-            activityTicket: insertedActivityTicket1,
-            activitySessionType: "offline",
-          });
-  
-        const jwtSecret = application
-          .get<ConfigService<{ HASURA_JWT_SECRET: string }>>(ConfigService)
-          .getOrThrow("HASURA_JWT_SECRET");
-  
-        const token = jwt.sign(
-          {
-            memberId: insertedMember.id,
-            permissions: [],
-          },
-          jwtSecret,
-        );
-  
-        const dto: FetchMemberRightActivityTicketDTO = {
-          activityTicketId: insertedActivityTicket1.id,
-          sessionId: insertedActivitySession2.id,
-        };
-  
-        const { body: data } = await request(application.getHttpServer())
-          .get(
-            `/activity/member_right?&activityTicketId=${dto.activityTicketId}${dto.sessionId ? `&sessionId=${dto.sessionId}` : ""}`,
-          )
-          .set("host", appHost.host)
-          .set("Authorization", `Bearer ${token}`)
-          .expect(200);
-  
-        expect(data.sessions.length).toEqual(1);
-        expect(data.sessions[0].id).toEqual(insertedActivitySession2.id);
-      });
-    });
-  
-    describe("Negative Testing with Valid Input", () => {
-      let insertedMember;
-      let insertedUnregisterMember;
-      let insertedActivity;
-      let insertedCategory;
-      let insertedActivityCategory;
-      let insertedActivitySession1;
-      let insertedActivityTicket1;
-      let insertedActivitySessionTicket1;
-      let insertedOrderLog;
-      let insertedProduct;
-      let insertedCurrency;
-      let insertedOrderProduct;
-  
-      beforeEach(async () => {
-        insertedMember = await createTestMember(manager, {
-          appId: app.id,
-          role: "app-owner",
-        });
-  
-        insertedUnregisterMember = await createTestMember(manager, {
-          appId: app.id,
-          role: "general-member",
-        });
-  
-        insertedActivity = await createTestActivity(manager, {
-          app: app,
-          organizer: insertedMember,
-          isPrivate: false, // scenario: 'holding' condition
-          publishedAt: new Date(), // scenario: 'holding' condition
-        });
-  
-        insertedCategory = await createTestCategory(manager, {
-          appId: app.id,
-          class: "activity",
-        });
-  
-        insertedActivityCategory = await createTestActivityCategory(manager, {
-          activity: insertedActivity,
-          category: insertedCategory,
-        });
-  
-        insertedActivitySession1 = await createTestActivitySession(manager, {
-          activity: insertedActivity,
-          startedAt: new Date("2020-01-01T00:00:00Z"),
-          endedAt: new Date("2020-01-02T00:00:00Z"),
-        });
-  
-        insertedActivityTicket1 = await createTestActivityTicket(manager, {
-          activity: insertedActivity,
-          startedAt: new Date("2020-01-01T00:00:00Z"),
-          endedAt: new Date("2020-01-02T00:00:00Z"),
-        });
-  
-        insertedActivitySessionTicket1 = await createTestActivitySessionTicket(
-          manager,
-          {
-            activitySession: insertedActivitySession1,
-            activityTicket: insertedActivityTicket1,
-            activitySessionType: "offline",
-          },
-        );
-  
-        insertedOrderLog = await createTestOrderLog(manager, {
-          member: insertedMember,
-          appId: app.id,
-        });
-  
-        insertedProduct = await createTestProduct(manager, {
-          id: `ActivityTicket_${insertedActivityTicket1.id}`,
-          type: "ActivityTicket",
-          target: insertedActivityTicket1.id,
-        });
-  
-        insertedCurrency = await createTestCurrency(manager, {
-          id: "TWD",
-        });
-  
-        insertedOrderProduct = await createTestOrderProduct(manager, {
-          order: insertedOrderLog,
-          product: insertedProduct,
-          currency: insertedCurrency,
-          productId: insertedActivity.id,
-          options: {
-            from: `/activities/${insertedActivity.id}`,
-            currencyId: insertedCurrency.id,
-            currencyPrice: 2000,
-          },
-        });
-      });
-  
-      it("should return an error when the member does not have access to the activity ticket", async () => {
-        const logSpy = jest.spyOn(console, 'error');
-
-        const jwtSecret = application
-          .get<ConfigService<{ HASURA_JWT_SECRET: string }>>(ConfigService)
-          .getOrThrow("HASURA_JWT_SECRET");
-  
-        const token = jwt.sign(
-          {
-            memberId: insertedUnregisterMember.id,
-            permissions: [],
-          },
-          jwtSecret,
-        );
-  
-        const dto: FetchMemberRightActivityTicketDTO = {
-          activityTicketId: insertedActivityTicket1.id,
-          sessionId: "",
-        };
-  
-        const { body: res } = await request(application.getHttpServer())
-          .get(
-            `/activity/member_right?&activityTicketId=${dto.activityTicketId}${dto.sessionId ? `&sessionId=${dto.sessionId}` : ""}`,
-          )
-          .set("host", appHost.host)
-          .set("Authorization", `Bearer ${token}`)
-          .expect(400);
-  
-        expect(res.code).toEqual("E_NOT_FOUND");
-  
-        expect(res.message).toEqual(
-          `Activity ticket data not found, activity_ticket_id: ${dto.activityTicketId}, member_id: ${insertedUnregisterMember.id}, session_id undefined`,
-        );
-
-        expect(logSpy).toHaveBeenCalled();
-        expect(logSpy).toBeCalledTimes(1);
-        logSpy.mockRestore();
-      });
-    });
-  
-    describe("Negative Testing with Invalid Input", () => {
-      it("should return a validation error for missing or invalid parameters)", async () => {
-        const jwtSecret = application
-          .get<ConfigService<{ HASURA_JWT_SECRET: string }>>(ConfigService)
-          .getOrThrow("HASURA_JWT_SECRET");
-  
-        const token = jwt.sign(
-          {
-            memberId: "invoke_member_id",
-            permissions: [],
-          },
-          jwtSecret,
-        );
-  
-        const dto: FetchMemberRightActivityTicketDTO = {
-          activityTicketId: "non_uuid",
-          sessionId: "non_uuid",
-        };
-  
-        const { body: res } = await request(application.getHttpServer())
-          .get(
-            `/activity/member_right?&activityTicketId=${dto.activityTicketId}${dto.sessionId ? `&sessionId=${dto.sessionId}` : ""}`,
-          )
-          .set("host", appHost.host)
-          .set("Authorization", `Bearer ${token}`)
-          .expect(400);
-  
-        expect(res.statusCode).toEqual(400);
-        expect(res.message).toEqual([
-          "activityTicketId must be a UUID",
-          "sessionId must be a UUID",
-        ]);
-      });
-    });
-  
-    describe("Permission Tests", () => {
-      describe("Executing API Calls with Different Permission Levels", () => {
-        it("should return an unauthorized error for users without the necessary permissions", async () => {
-          const dto: FetchMemberRightActivityTicketDTO = {
-            activityTicketId: "non_uuid",
-            sessionId: "non_uuid",
-          };
-  
-          const { body: res } = await request(application.getHttpServer())
-            .get(
-              `/activity/member_right?&activityTicketId=${dto.activityTicketId}${dto.sessionId ? `&sessionId=${dto.sessionId}` : ""}`,
-            )
-            .set("host", appHost.host)
-            .expect(401);
-  
-        });
-      });
-    });
-  });
-
 });
